move 'activation' module contents under 'mode'
