+++ tests/unit/test_core/test_activationresult.py
