+++ tests/unit/test_core/test_mode/test_mode.py
